+++ postrify-frontend/src/app/app.routes.ts
@@ -1,10 +1,10 @@
 import { Routes } from '@angular/router';
-import { LoginComponent } from './login/login.component';
-import { RegisterComponent } from './register/register.component';
-import { HomeComponent } from './home/home.component';
-import { PageNotFoundComponentComponent } from './page-not-found-component/page-not-found-component.component';
-import { PostFormComponent } from './post-form/post-form.component';
-import { PostDetailComponent } from './post-detail/post-detail.component';
+import { LoginComponent } from './components/login/login.component';
+import { RegisterComponent } from './components/register/register.component';
+import { HomeComponent } from './components/home/home.component';
+import { PageNotFoundComponentComponent } from './components/page-not-found-component/page-not-found-component.component';
+import { PostFormComponent } from './components/post-form/post-form.component';
+import { PostDetailComponent } from './components/post-detail/post-detail.component';
 
 export const routes: Routes = [
   { path: '', component: HomeComponent },
