+++ postrify-frontend/src/app/components/footer/footer.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import packageJson from '../../../package.json';
+import packageJson from '../../../../package.json';
 
 @Component({
   selector: 'app-footer',
