+++ postrify-frontend/src/app/components/post-detail/post-detail.component.ts
