+++ postrify-frontend/src/app/components/toast/toast.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { ToastService, Toast } from '../services/toast.service';
+import { ToastService, Toast } from '../../services/toast.service';
 import { Subscription } from 'rxjs';
 import { CommonModule } from '@angular/common';
